remote: https://github.com/ARM-Software/devlib.git

+++ external/devlib/devlib/instrument/daq.py
@@ -16,8 +16,10 @@
 import os
 import shutil
 import tempfile
-from itertools import chain
+import time
+from itertools import chain, zip_longest
 
+from devlib.host import PACKAGE_BIN_DIRECTORY
 from devlib.instrument import Instrument, MeasurementsCsv, CONTINUOUS
 from devlib.exception import HostError
 from devlib.utils.csvutil import csvwriter, create_reader
@@ -45,7 +47,8 @@ class DaqInstrument(Instrument):
                  dv_range=0.2,
                  sample_rate_hz=10000,
                  channel_map=(0, 1, 2, 3, 4, 5, 6, 7, 16, 17, 18, 19, 20, 21, 22, 23),
-                 keep_raw=False
+                 keep_raw=False,
+                 time_as_clock_monotonic=True
                  ):
         # pylint: disable=no-member
         super(DaqInstrument, self).__init__(target)
@@ -53,6 +56,7 @@ class DaqInstrument(Instrument):
         self._need_reset = True
         self._raw_files = []
         self.tempdir = None
+        self.target_monotonic_clock_at_start = 0.0
         if DaqClient is None:
             raise HostError('Could not import "daqpower": {}'.format(import_error_mesg))
         if labels is None:
@@ -76,10 +80,29 @@ class DaqInstrument(Instrument):
                                                  channel_map=channel_map,
                                                  labels=labels)
         self.sample_rate_hz = sample_rate_hz
+        self.time_as_clock_monotonic = time_as_clock_monotonic
 
+        self.add_channel('Time', 'time')
         for label in labels:
             for kind in ['power', 'voltage']:
                 self.add_channel(label, kind)
+
+        if time_as_clock_monotonic:
+            host_path = os.path.join(PACKAGE_BIN_DIRECTORY, self.target.abi,
+                                     'get_clock_monotonic')
+            self.clock_monotonic_cmd = self.target.install_if_needed(host_path,
+                                                                     search_system_binaries=False)
+
+    def calculate_monotonic_offset(self):
+        time_before = time.time()
+        out = self.target.execute(self.clock_monotonic_cmd)
+        time_after = time.time()
+
+        remote_clock_monotonic = float(out)
+        propagation_delay = (time_after - time_before) / 2
+        monotonic_at_end = remote_clock_monotonic + propagation_delay
+
+        return time_after - monotonic_at_end
 
     def reset(self, sites=None, kinds=None, channels=None):
         super(DaqInstrument, self).reset(sites, kinds, channels)
@@ -90,8 +113,18 @@ class DaqInstrument(Instrument):
 
     def start(self):
         if self._need_reset:
-            self.reset()
+            # Preserve channel order
+            self.reset(channels=self.channels.keys())
+
+        if self.time_as_clock_monotonic:
+            target_monotonic_offset = self.calculate_monotonic_offset()
+            time_start = time.time()
+
         self.daq_client.start()
+
+        if self.time_as_clock_monotonic:
+            time_end = time.time()
+            self.target_monotonic_clock_at_start = (time_start + time_end) / 2 - target_monotonic_offset
 
     def stop(self):
         self.daq_client.stop()
@@ -118,32 +151,32 @@ class DaqInstrument(Instrument):
                     site_readers[site] = reader
                     file_handles.append(fh)
                 except KeyError:
-                    message = 'Could not get DAQ trace for {}; Obtained traces are in {}'
-                    raise HostError(message.format(site, self.tempdir))
+                    if not site.startswith("Time"):
+                        message = 'Could not get DAQ trace for {}; Obtained traces are in {}'
+                        raise HostError(message.format(site, self.tempdir))
 
             # The first row is the headers
-            channel_order = []
+            channel_order = ['Time_time']
             for site, reader in site_readers.items():
                 channel_order.extend(['{}_{}'.format(site, kind)
                                       for kind in next(reader)])
 
-            def _read_next_rows():
-                parts = []
-                for reader in site_readers.values():
-                    try:
-                        parts.extend(next(reader))
-                    except StopIteration:
-                        parts.extend([None, None])
-                return list(chain(parts))
+            def _read_rows():
+                row_iter = zip_longest(*site_readers.values(), fillvalue=(None, None))
+                for raw_row in row_iter:
+                    raw_row = list(chain.from_iterable(raw_row))
+                    raw_row.insert(0, _read_rows.row_time_s)
+                    yield raw_row
+                    _read_rows.row_time_s += 1.0 / self.sample_rate_hz
+
+            _read_rows.row_time_s = self.target_monotonic_clock_at_start
 
             with csvwriter(outfile) as writer:
                 field_names = [c.label for c in self.active_channels]
                 writer.writerow(field_names)
-                raw_row = _read_next_rows()
-                while any(raw_row):
+                for raw_row in _read_rows():
                     row = [raw_row[channel_order.index(f)] for f in field_names]
                     writer.writerow(row)
-                    raw_row = _read_next_rows()
 
             return MeasurementsCsv(outfile, self.active_channels, self.sample_rate_hz)
         finally:
@@ -156,5 +189,5 @@ class DaqInstrument(Instrument):
     def teardown(self):
         self.daq_client.close()
         if not self.keep_raw:
-            if os.path.isdir(self.tempdir):
+            if self.tempdir and os.path.isdir(self.tempdir):
                 shutil.rmtree(self.tempdir)
